Refactored Header component structure
