+++ client/public/css/styles.css
@@ -83,7 +83,7 @@ header nav {
   -webkit-transition: height 0.7s;
   transition: height 0.7s;
 }
-header nav > ul {
+.nav-primary-container {
   padding: 0;
   margin: 0;
   list-style-type: none;
@@ -96,12 +96,12 @@ header nav > ul {
           justify-content: space-between;
   background: var(--color-dark-blue);
 }
-header nav > ul > li {
+header .nav-item-container {
   padding: 0.5em 1em;
   width: 50%;
   display: block;
 }
-header nav > ul > li .submenu {
+header .nav-item-container .submenu {
   font-family: 'PT Sans', sans-serif;
   font-size: 1.25em;
   color: #fff;
@@ -120,7 +120,7 @@ header nav > ul > li .submenu {
           transform: scale(3) translateY(50px);
   visibility: hidden;
 }
-header nav > ul > li .submenu > div {
+header .nav-item-container .submenu > div {
   margin: 0.5em;
 }
 header .link-list-container ul {
