+++ client/src/components/Header/AboutList/aboutItems.js
@@ -1,0 +1,18 @@
+export const aboutItems = [
+  {
+    link: "#",
+    text: "About the Show"
+  },
+  {
+    link: "#",
+    text: "News"
+  },
+  {
+    link: "#",
+    text: "Sign Up"
+  },
+  {
+    link: "#",
+    text: "Support the Show"
+  }
+];
