+++ client/src/components/Header/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Branding from '../Branding';
-import Navbar from '../Navbar';
+import Branding from './Branding';
+import Navbar from './Navbar';
 
 
 const Header = () => {
