--- client/src/components/SubscriptionListItem/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const SubscriptionListItem = ({service}) => {
-  return (
-    <li>
-      <a href={service.link} title={service.title} target="_blank">
-        <img src={service.image} alt={service.alt} />
-      </a>
-    </li>
-  );
-}
-
-export default SubscriptionListItem;
